+++ NEWS.md
@@ -1,3 +1,13 @@
+# spant 0.3.0
+* Added stackplot method for fit objects.
+* Added functions for registering and visualising SVS volumes on images and 
+performing partial volume correction.
+* Philips "list data" also now reads noise scans.
+* calc_coil_noise_cor, calc_coil_noise_sd functions added to aid coil 
+combination.
+* Documentation updates for plotting methods.
+* Added some simulation methods to userland.
+
 # spant 0.2.0
 * Added Siemens RDA format reader.
 * Added Philips "list data" format reader.